tests/integration/test_rate_limiting.py
"""
Rate limiting and error recovery integration tests.

Tests that the system properly handles:
- Google API rate limits and quota errors
- Network failures and timeouts
- Retry logic with exponential backoff
- Circuit breaker behavior
- Graceful degradation

These tests simulate error conditions and verify recovery mechanisms.
"""

import time
from unittest.mock import Mock

import pytest
from googleapiclient.errors import HttpError

from src.services import GoogleSheetsService, RetryHandler
from src.services.error_classifier import ErrorClassifier, ErrorType


@pytest.mark.integration
@pytest.mark.api
@pytest.mark.slow
class TestRateLimitingBehavior:
    """Test system behavior when hitting rate limits."""

    def test_retry_on_rate_limit_error(
        self,
        real_sheets_service: GoogleSheetsService,
    ):
        """Test that rate limit errors trigger retry logic."""
        # This test verifies the retry handler is configured correctly
        # Actual rate limiting errors are hard to simulate without hitting real limits

        retry_handler = RetryHandler(max_retries=3)

        # Verify retry handler configuration
        assert retry_handler.max_retries == 3
        assert retry_handler.base_delay > 0

    def test_exponential_backoff_timing(
        self,
    ):
        """Test that retry delays follow exponential backoff pattern."""
        retry_handler = RetryHandler(max_retries=4, base_delay=0.1)

        # Calculate expected delays
        delays = []
        for attempt in range(4):
            delay = retry_handler._calculate_delay(attempt)
            delays.append(delay)

        # Verify exponential growth pattern (with jitter tolerance)
        # Each delay should be roughly 2x the previous (within jitter variance)
        for i in range(1, len(delays)):
            # Delays should generally increase
            # (exact 2x not guaranteed due to jitter)
            assert (
                delays[i] > delays[i - 1] * 0.5
            ), "Delays should increase exponentially"

        print(f"\nRetry delays (with jitter): {[f'{d:.3f}s' for d in delays]}")

    def test_max_retries_limit(
        self,
    ):
        """Test that retries stop after max_retries is reached."""
        retry_handler = RetryHandler(max_retries=3)

        # Simulate function that always fails
        call_count = 0

        @retry_handler.with_retry
        def always_fails():
            nonlocal call_count
            call_count += 1
            raise Exception("Test error")

        # Should fail after max_retries + 1 attempts (initial + 3 retries)
        with pytest.raises(Exception):
            always_fails()

        assert call_count == 4, f"Expected 4 attempts (1 + 3 retries), got {call_count}"

    def test_successful_retry_recovery(
        self,
    ):
        """Test that operation succeeds after transient error."""
        retry_handler = RetryHandler(max_retries=3, base_delay=0.01)

        call_count = 0

        @retry_handler.with_retry
        def fails_twice_then_succeeds():
            nonlocal call_count
            call_count += 1
            if call_count < 3:
                raise Exception("Transient error")
            return "success"

        result = fails_twice_then_succeeds()

        assert result == "success"
        assert call_count == 3, "Should have called 3 times before success"


@pytest.mark.integration
@pytest.mark.api
class TestErrorClassification:
    """Test error classification and appropriate handling."""

    def test_classify_rate_limit_error(
        self,
    ):
        """Test classification of rate limit errors."""
        classifier = ErrorClassifier()

        # Simulate rate limit error (429)
        error = HttpError(
            resp=Mock(status=429, reason="Too Many Requests"),
            content=b"Rate limit exceeded",
        )

        error_type = classifier.classify_error(error)

        assert error_type == ErrorType.RATE_LIMIT
        assert classifier.is_retryable(error)

    def test_classify_server_error(
        self,
    ):
        """Test classification of server errors (5xx)."""
        classifier = ErrorClassifier()

        # Simulate server error (503)
        error = HttpError(
            resp=Mock(status=503, reason="Service Unavailable"),
            content=b"Service unavailable",
        )

        error_type = classifier.classify_error(error)

        assert error_type == ErrorType.SERVER_ERROR
        assert classifier.is_retryable(error)

    def test_classify_permission_error(
        self,
    ):
        """Test classification of permission errors (not retryable)."""
        classifier = ErrorClassifier()

        # Simulate permission error (403)
        error = HttpError(
            resp=Mock(status=403, reason="Forbidden"), content=b"Permission denied"
        )

        error_type = classifier.classify_error(error)

        assert error_type == ErrorType.PERMISSION_ERROR
        assert not classifier.is_retryable(error)

    def test_classify_not_found_error(
        self,
    ):
        """Test classification of not found errors (not retryable)."""
        classifier = ErrorClassifier()

        # Simulate not found error (404)
        error = HttpError(
            resp=Mock(status=404, reason="Not Found"), content=b"File not found"
        )

        error_type = classifier.classify_error(error)

        assert error_type == ErrorType.NOT_FOUND
        assert not classifier.is_retryable(error)


@pytest.mark.integration
@pytest.mark.api
@pytest.mark.slow
class TestErrorRecovery:
    """Test error recovery mechanisms."""

    def test_graceful_degradation_on_api_failure(
        self,
        real_sheets_service: GoogleSheetsService,
    ):
        """Test that system handles API failures gracefully."""
        # Try to access an invalid spreadsheet
        with pytest.raises(Exception) as exc_info:
            real_sheets_service.read_sheet_data(
                spreadsheet_id="invalid_id", range_name="Sheet1!A1:B2"
            )

        # Should raise a clear error, not crash
        assert exc_info.value is not None

    def test_partial_data_handling(
        self,
        real_sheets_service: GoogleSheetsService,
        test_spreadsheet_id: str,
    ):
        """Test handling of partial/incomplete data."""
        # Write partial data (missing some expected columns)
        partial_data = [
            ["Date", "Project"],  # Missing other columns
            ["2024-01-01", "PROJECT_A"],
        ]

        real_sheets_service.update_sheet_data(
            spreadsheet_id=test_spreadsheet_id,
            range_name="Sheet1!A1",
            values=partial_data,
        )

        # Read should succeed, even with partial data
        data = real_sheets_service.read_sheet_data(
            spreadsheet_id=test_spreadsheet_id, range_name="Sheet1!A1:H10"
        )

        assert len(data) > 0

    def test_empty_response_handling(
        self,
        real_sheets_service: GoogleSheetsService,
        test_spreadsheet_id: str,
    ):
        """Test handling of empty API responses."""
        # Try to read from empty range
        data = real_sheets_service.read_sheet_data(
            spreadsheet_id=test_spreadsheet_id, range_name="Sheet1!Z100:Z200"
        )

        # Should return empty list, not error
        assert isinstance(data, list)
        assert len(data) == 0


@pytest.mark.integration
@pytest.mark.api
@pytest.mark.slow
class TestNetworkResilience:
    """Test resilience to network issues."""

    def test_timeout_handling(
        self,
    ):
        """Test that timeouts are properly handled."""
        retry_handler = RetryHandler(max_retries=2, timeout=0.01)  # Very short timeout

        @retry_handler.with_retry
        def slow_operation():
            time.sleep(1.0)  # Slower than timeout
            return "success"

        # Should fail due to timeout
        with pytest.raises(Exception):
            slow_operation()

    def test_connection_error_retry(
        self,
    ):
        """Test retry behavior on connection errors."""
        retry_handler = RetryHandler(max_retries=3, base_delay=0.01)

        call_count = 0

        @retry_handler.with_retry
        def connection_error_then_success():
            nonlocal call_count
            call_count += 1
            if call_count < 2:
                raise ConnectionError("Network unavailable")
            return "connected"

        result = connection_error_then_success()

        assert result == "connected"
        assert call_count == 2


@pytest.mark.integration
@pytest.mark.api
@pytest.mark.slow
class TestCircuitBreakerBehavior:
    """Test circuit breaker pattern (if implemented)."""

    def test_circuit_breaker_opens_after_failures(
        self,
    ):
        """Test that circuit breaker opens after repeated failures."""
        # Note: This test assumes circuit breaker is implemented
        # If not yet implemented, this can be a placeholder for future work

        pytest.skip("Circuit breaker not yet implemented - planned for future")

    def test_circuit_breaker_resets_after_success(
        self,
    ):
        """Test that circuit breaker resets after successful operation."""
        pytest.skip("Circuit breaker not yet implemented - planned for future")


@pytest.mark.integration
@pytest.mark.api
class TestBatchOperationResilience:
    """Test resilience of batch operations."""

    def test_partial_batch_success(
        self,
        real_sheets_service: GoogleSheetsService,
        test_spreadsheet_id: str,
    ):
        """Test that batch operations handle partial failures gracefully."""
        # Write test data to multiple cells
        # If one write fails, others should still succeed

        # This is a placeholder - full implementation would test actual batch API calls
        test_data = [["A", "B"], ["1", "2"]]

        # Should succeed for valid operations
        real_sheets_service.update_sheet_data(
            spreadsheet_id=test_spreadsheet_id,
            range_name="Sheet1!A1",
            values=test_data,
        )

        # Verify data was written
        read_data = real_sheets_service.read_sheet_data(
            spreadsheet_id=test_spreadsheet_id, range_name="Sheet1!A1:B2"
        )

        assert read_data == test_data


@pytest.mark.integration
@pytest.mark.api
@pytest.mark.slow
class TestLongRunningOperations:
    """Test handling of long-running operations."""

    def test_long_running_aggregation_timeout(
        self,
    ):
        """Test that long-running aggregations don't timeout prematurely."""
        # This would test the complete aggregation pipeline with timeout handling
        # Placeholder for now

        pytest.skip("Long-running operation test requires full pipeline setup")

    def test_progress_tracking_during_long_operation(
        self,
    ):
        """Test that progress is tracked during long-running operations."""
        # Verify progress tracking mechanisms work
        pytest.skip("Progress tracking test requires full pipeline setup")
