requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0
pytest-benchmark>=4.0.0
memory-profiler>=0.61.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
types-requests>=2.31.0

# Development Tools
pre-commit>=3.3.0
ipdb>=0.13.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
